+++ functions/f-bw_cron.sh
@@ -34,5 +34,7 @@ function f-bw_cron {
 #catcherror
 	bw_log "2" "Finished ${bw_action}"
 
+	bw_sendmail "2" "SUCCESS" "Successfully Finished ${bw_action} of ${bw_project}"
+
 	return 0	
 }